+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.x] -
+## [0.1.2] -
 ### Added
-- Added development guide
+- Added development guide and storybook integration
 - Added new story for UsersTable in a Drawer
+- Support mutually exclusive roles. In case set, only 1 role would be allowed to be assigned.
 ### Fixed
 - Fixed issues when running Storybook locally
+- Fixed issues preventing correct render when props are being updated
 
 ## [0.1.1] - 2018-08-14
 ### Fixed